ft_ls ok dans l'autre dossier que le dossier courrant, printf toujours random bug

+++ printf/includes/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: apoisson <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/12/10 01:22:47 by apoisson          #+#    #+#             */
-/*   Updated: 2017/03/12 03:28:12 by apoisson         ###   ########.fr       */
+/*   Updated: 2017/03/25 04:29:26 by apoisson         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -151,7 +151,7 @@ typedef struct		s_data
 	char			*buff;
 	char			*format;
 	va_list			ap;
-	size_t			len;
+	int			len;
 	t_conv			*conv;
 	t_arg			*arg;
 	t_uni			*uni;
